# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/todo/lists": {
      "get": {
        "tags": [
          "todo-controller-impl"
        ],
        "summary": "Get all lists",
        "description": "Get a list of all todo lists",
        "operationId": "getLists",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "uniqueItems": true,
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoListDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "todo-controller-impl"
        ],
        "description": "Add a new todo list",
        "operationId": "createList",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoListDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/todo/lists/{name}": {
      "get": {
        "tags": [
          "todo-controller-impl"
        ],
        "description": "Get a list of todo items with the specified name",
        "operationId": "getList",
        "parameters": [
          {
            "name": "listId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TodoListDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "todo-controller-impl"
        ],
        "operationId": "updateList",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoListDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "todo-controller-impl"
        ],
        "operationId": "deleteList",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/todo/lists/{name}/items": {
      "get": {
        "tags": [
          "todo-controller-impl"
        ],
        "operationId": "getItems",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "uniqueItems": true,
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoEntryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "todo-controller-impl"
        ],
        "operationId": "createItem",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoEntryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/todo/lists/{name}/items/{itemName}": {
      "get": {
        "tags": [
          "todo-controller-impl"
        ],
        "operationId": "getItem",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TodoEntryDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "todo-controller-impl"
        ],
        "operationId": "updateItem",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoEntryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "todo-controller-impl"
        ],
        "operationId": "deleteItem",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/todo/lists/{name}/itemNames": {
      "get": {
        "tags": [
          "todo-controller-impl"
        ],
        "operationId": "getItemNames",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "uniqueItems": true,
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/version": {
      "get": {
        "tags": [
          "ApiInfo"
        ],
        "summary": "Get API version",
        "description": "Get API version",
        "operationId": "getVersion",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VersionDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TodoListDto": {
        "required": [
          "color",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 64,
            "minLength": 2,
            "type": "string"
          },
          "description": {
            "maxLength": 2048,
            "minLength": 0,
            "type": "string"
          },
          "color": {
            "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
            "type": "string"
          }
        }
      },
      "TodoEntryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "done": {
            "type": "boolean"
          },
          "dueTo": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "VersionDto": {
        "type": "object",
        "properties": {
          "version": {
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-((?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\\.(?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\\+([0-9a-zA-Z-]+(?:\\.[0-9a-zA-Z-]+)*))?$",
            "type": "string"
          },
          "stable": {
            "type": "boolean"
          }
        }
      }
    }
  }
}